CTF: Fix map entities being missing or placed incorrectly
The map "agctf_legacy" has trigger some issues with the way I was reading custom entities, causing weapons and items being misplaced and the flags being missing.

The reason of the issue was that I spawn the custom entity when the "origin" key is matched first. But in this map, the "angles" key was first, setting properties to some unknown entity.

This new way should be more simpler and safer, by being independent of the order of the keys.

Thanks to Safety1st for reporting the issue.