Fix BHL not working on Linux in some cases
BugfixedHL wasn't working on Linux for some reason when using Metamod from https://github.com/Bots-United/metamod-p/

This metamod didn't find the game dll so I switch the name of hl.so to hl_i386.so and it worked but that leaded to another issue.

When a hl.so file is already present (like the one from vanilla HL), Metamod will not look for my hl_i386.so, so AG Mod X will not work.

I managed to get a mirror of the previous Metamod I was using, which was deleted by his author.

Let's hope this fix it.